moved functions out of memory to MenuParsers.py to clean up the clutter
